+++ docs/test/maps.js
@@ -4,34 +4,53 @@ maps = {
   small: {
     name: 'Mitochondia [69,034 bp]',
     type: 'basic',
-    url: 'maps/mito.json',
-  },
-  linear_1_6: {
-    name: 'Linear Plasmid [90,229 bp]',
-    type: 'basic',
-    url: 'maps/linear.1.6.json',
+    url: 'maps/basic_mito.json',
   },
   small_noplots: {
     name: 'Mitochondia No Plots [69,034 bp]',
     type: 'basic',
-    url: 'maps/mito_no_plots.json',
+    url: 'maps/basic_mito_no_plots.json',
   },
   medium: {
     name: 'E. coli [4,641,652 bp]',
     type: 'basic',
-    url: 'maps/ecoli.json',
+    url: 'maps/basic_ecoli.json',
+  },
+
+  // CONTIGS
+  small_contigs: {
+    name: 'Contigs: MNHM01000001 [133,342 bp]',
+    type: 'contigs',
+    url: 'maps/contigs_small.json',
+  },
+  ecoli_pa2_contigs: {
+    name: 'Contigs: Ecoli PA2 [5,246,924 bp]',
+    type: 'contigs',
+    url: 'maps/contigs_ecoli_pa2.json',
+  },
+
+  // LARGE - NOT WORKING
+  large: {
+    name: 'L. guizhouensis [9,997,872 bp]',
+    type: 'large',
+    url: 'maps/large_lentzea.json',
+  },
+  large_noplots: {
+    name: 'L. guizhouensis No Plots [9,997,872 bp]',
+    type: 'large',
+    url: 'maps/large_lentzea_no_plots.json',
   },
 
   // TESTING
   locations: {
     name: 'Locations Test [69,034 bp]',
     type: 'test',
-    url: 'maps/locations.json',
+    url: 'maps/test_locations.json',
   },
   blast: {
     name: 'Blast Overlap Test [69,034 bp]',
     type: 'test',
-    url: 'maps/blast_overlap.json',
+    url: 'maps/test_blast_overlap.json',
   },
   test_single_bases: {
     name: 'Test Single Base Features',
@@ -43,73 +62,49 @@ maps = {
   labels: {
     name: 'Label Test [69,034 bp]',
     type: 'labels',
-    url: 'maps/labels.json',
+    url: 'maps/labels_1.json',
   },
   labels2: {
     name: 'Label 2 Test [69,034 bp]',
     type: 'labels',
-    url: 'maps/labels2.json',
+    url: 'maps/labels_2.json',
   },
   labels3: {
     name: 'Label 3 Test [69,034 bp]',
     type: 'labels',
-    url: 'maps/labels3.json',
+    url: 'maps/labels_3.json',
   },
   pcDNA3: {
     name: 'pcDNA3.1',
     type: 'labels',
-    url: 'maps/labels-pcDNA3.1.json',
+    url: 'maps/labels_pcDNA3.1.json',
   },
   pcET30c: {
     name: 'pET-30c',
     type: 'labels',
-    url: 'maps/labels-pET-30c.json',
+    url: 'maps/labels_pET-30c.json',
   },
   pPICZ: {
     name: 'pPICZ',
     type: 'labels',
-    url: 'maps/labels-pPICZ.json',
+    url: 'maps/labels_pPICZ.json',
   },
   phastest: {
     name: 'phastest example',
     type: 'labels',
-    url: 'maps/phastest2.json',
-  },
-
-  // CONTIGS
-  small_contigs: {
-    name: 'Contigs: MNHM01000001 [133,342 bp]',
-    type: 'contigs',
-    url: 'maps/small_contigs.json',
-  },
-  ecoli_pa2_contigs: {
-    name: 'Contigs: Ecoli PA2 [5,246,924 bp]',
-    type: 'contigs',
-    url: 'maps/ecoli_pa2_contigs.json',
-  },
-
-  // LARGE
-  large: {
-    name: 'L. guizhouensis [9,997,872 bp]',
-    type: 'large',
-    url: 'maps/lentzea.json',
-  },
-  large_noplots: {
-    name: 'L. guizhouensis No Plots [9,997,872 bp]',
-    type: 'large',
-    url: 'maps/lentzea_noplots.json',
-  },
-  blast_big: {
-    name: 'Blast BIG',
-    type: 'large',
-    url: 'maps/blast_big.json',
+    url: 'maps/labels_phastest2.json',
   },
 
   // VERSION
   version_0_1: {
     name: 'Mitochondia [69,034 bp]; Version 0.1',
     type: 'version',
-    url: 'maps/mito-0.1.json',
+    url: 'maps/version_mito-0.1.json',
+  },
+  linear_1_6: {
+    name: 'Linear Plasmid [90,229 bp]',
+    type: 'version',
+    url: 'maps/version_linear.1.6.json',
   },
 
   // BAD